+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/Methods_build.cs
@@ -168,7 +168,7 @@ public static partial class SessionBaseExtensionMethods
         }
         finally
         {
-            session.RestoreDisplayCompleteQuery();
+            session.RestoreMainDisplay();
         }
     }
 }